Remove timing information from JI login
Removes timing information from two functions that are executed during the JI
login process.

In `valid_password()` there was a timing-variable, byte-by-byte comparison of
the salted scrypt hash of submitted password with the salted scrypt hash of
their actual password. Since the salt is kept secret from the Journalist, this
makes it impossible to know what the final value of the submitted password hash
is. Thus an attacker may learn how many prefix bytes in common the hash of their
submitted password has with actual password hash, but not what those bytes are.
Due to the avalanche effect of cryptographic hash functions, this makes this
information totally useless. Therefore, *this timing information is not
exploitable*, but we are removing it anyway because we have gotten several
complaints about it, and would rather not go through the explanation of why it's
safe here again.

In `verify_token()` there was a timing-variable, byte-by-byte comparison of
`self.last_token` (a Journalist field corresponding to the last token a
Journalist attempted to login with) with the submitted token. If
`self.last_token` was a valid login token, and assuming best possible
circumstances (client time was skewed ahead of server and attacker immediately
starts guessing after it's entered, giving a >60s window of validity, timing
information resolution is perfect, etc.) it can be exploited to increase the
probability of an attacker guessing a correct token by ~2.2 times during that
period:

```
>>> p = (10 / float(10 ** 6))
>>> (sum([9 / (float(10 ** (6 - n)) * float(10 ** n)) for n in range(5)]) + (2 * p)) / (3 * p)
2.1666666666666665
```

Still, when you look at the actual value, it's pretty insignificant:

```
>>> sum([9 / (float(10 ** (6 - n)) * float(10 ** n)) for n in range(5)]) + (2 * p)
6.500000000000001e-05
```

Although the information gain is almost nothing here even under the best-case
scenario, we're again going to patch this one with a constant-time comparison
just to avoid bug reports in the future, and play it extra safe.

Credit to @EdOverflow for pointing out the timing leak in `verify_token()`.