Creates handler for reboot-and-wait-for
In pursuit of a DRY config, let's declare one handler for reboots and
use that where necessary to reboot (and wait for host to come back up)
via the same consistent logic. The handler is declared once within the
"common" role, and is therefore available to all roles within the same
play.

The catch, then, is that any attempts to reference the handler from
within other (subsequent) plays will fail. That's a bit brittle, but
long-term we should rely more heavily on upstream to avoid managing
custom logic for simple provisioning tasks.